Refractor para el final del sprint

+++ src/components/Cuenta/AdministrarCorreo.js
@@ -1,14 +1,16 @@
-import React, { Component } from "react";
-import auth from "../../helpers/auth";
+import React, { Component } from 'react';
 import axios from 'axios';
 import swal from 'sweetalert';
+import auth from '../../helpers/auth';
+import { Redirect } from 'react-router-dom';
 
 export default class AdministrarCorreo extends Component {
     constructor(props) {
         super(props);
         this.state = {
             correo: '',
-            correos: []
+            correos: [],
+            redirect: false
         };
         this.handleInputChange = this.handleInputChange.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
@@ -20,12 +22,25 @@ export default class AdministrarCorreo extends Component {
     }
 
     getEmails() {
-        const cedula = auth.getInfo().cedula;
-        axios.get(`http://localhost:5000/correo/${cedula}`)
-            .then(res => {
-                this.setState({
-                    correos: res.data
-                });
+        auth.verifyToken()
+            .then(value => {
+                if (value) {
+                    const cedula = auth.getInfo().cedula;
+                    axios.get(`http://localhost:5000/correo/${cedula}`)
+                        .then(res => {
+                            if (res.data.success) {
+                                this.setState({
+                                    correos: res.data.emails
+                                });
+                            }
+                        })
+                        .catch((err) => console.log(err));
+                } else {
+                    this.setState({
+                        redirect: true
+                    });
+                    auth.logOut();
+                }
             })
             .catch((err) => console.log(err));
     }
@@ -40,33 +55,68 @@ export default class AdministrarCorreo extends Component {
 
     deleteEmail(e, email) {
         e.preventDefault();
-        swal({
-            title: "Confirmación",
-            text: `Se eliminará ${email}`,
-            icon: "warning",
-            buttons: ["Cancelar", "Confirmar"],
-            dangerMode: true,
-        })
-            .then((willDelete) => {
-                if (willDelete) {
-                    axios.delete('http://localhost:5000/correo', { data: { correo: email } })
-                        .then(() => {
-                            this.getEmails();
-                        })
-                        .catch((err) => console.log(err));
+        auth.verifyToken()
+            .then(value => {
+                if (value) {
+                    swal({
+                        title: "Confirmación",
+                        text: `Se eliminará ${email}`,
+                        icon: "warning",
+                        buttons: ["Cancelar", "Confirmar"],
+                        dangerMode: true,
+                    })
+                        .then((willDelete) => {
+                            if (willDelete) {
+                                axios.delete('http://localhost:5000/correo', { data: { correo: email } })
+                                    .then(() => {
+                                        this.getEmails();
+                                    })
+                                    .catch((err) => console.log(err));
+                            }
+                        });
+                } else {
+                    this.setState({
+                        redirect: true
+                    });
+                    auth.logOut();
                 }
-            });
+            })
+            .catch((err) => console.log(err));
     }
 
     handleSubmit(e) {
         e.preventDefault();
-        const cedula = auth.getInfo().cedula;
-        axios.post(`http://localhost:5000/correo/${cedula}`, { correo: this.state.correo })
-            .then(() => {
-                this.getEmails()
-                this.setState({
-                    correo: ''
-                })
+        auth.verifyToken()
+            .then(value => {
+                if (value) {
+                    axios.post('http://localhost:5000/correo/verificar_correo', { correo: this.state.correo })
+                        .then(res => {
+                            if (!res.data.taken) {
+                                const cedula = auth.getInfo().cedula;
+                                axios.post(`http://localhost:5000/correo/${cedula}`, { correo: this.state.correo })
+                                    .then(() => {
+                                        this.getEmails()
+                                        this.setState({
+                                            correo: ''
+                                        })
+                                    })
+                                    .catch((err) => console.log(err));
+                            } else {
+                                swal({
+                                    title: "Este correo ya existe",
+                                    text: "El correo que se digitó ya se encuentra registrado en el sistema.",
+                                    icon: "warning",
+                                    button: "Ok",
+                                });
+                            }
+                        })
+                        .catch((err) => console.log(err));
+                } else {
+                    this.setState({
+                        redirect: true
+                    });
+                    auth.logOut();
+                }
             })
             .catch((err) => console.log(err));
     }
@@ -86,7 +136,7 @@ export default class AdministrarCorreo extends Component {
     }
 
     render() {
-        return (
+        return (this.state.redirect ? <Redirect to='/' /> :
             <div className="container">
                 <h3 className="mb-4 text-center">Administración de Correos</h3>
                 <form onSubmit={this.handleSubmit}>
@@ -101,6 +151,8 @@ export default class AdministrarCorreo extends Component {
                         </div>
                     </div>
                 </form>
+                {this.state.correos === 0 &&
+                    <p className="my-muted">No tiene correos registrados</p>}
                 {this.emailsList()}
             </div>
         );
